% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_param}
\alias{add_param}
\title{Add element to the list (shortcut)}
\usage{
add_param(previous_list, item_to_add)
}
\arguments{
\item{previous_list}{List to add items to}

\item{item_to_add}{Item to add to list}
}
\description{
Add element to the list (shortcut)
}
