% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_path_to_graphs}
\alias{add_path_to_graphs}
\title{Add path to graphs}
\usage{
add_path_to_graphs(x)
}
\arguments{
\item{x}{text to convert}
}
\value{
updated text
}
\description{
By default pandoc rtf is run in its own location and does not locate the images properly
manual intervention is required to explicitly point the images
}
