% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate_quantitative_to_qualitative_materiality}
\alias{aggregate_quantitative_to_qualitative_materiality}
\title{Aggregate multiple numerical materiality exposures into a single, qualitative}
\usage{
aggregate_quantitative_to_qualitative_materiality(x)
}
\arguments{
\item{x}{a vector of numeric materialities}
}
\description{
The thresholds are defined so that 10 and above is "high",
anything between 5 (incl.) and 10 (excl.) is medium.
instead of 1000 I could use infinity (but up to at least 100 classes it makes no difference)
}
