% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alternative_sleep}
\alias{alternative_sleep}
\title{Replacement for Sys.sleep()}
\usage{
alternative_sleep(time)
}
\arguments{
\item{time}{Time to sleep (in seconds)}
}
\description{
Replacement for Sys.sleep()
}
