% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{capitalize}
\alias{capitalize}
\title{Make the first letter of a string upper case}
\usage{
capitalize(string)
}
\arguments{
\item{string}{the string to convert}
}
\description{
Make the first letter of a string upper case
}
