% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_add.R
\name{check_required_libraries}
\alias{check_required_libraries}
\title{Check if the files (e.g. javascript) required by the report are available in the main package}
\usage{
check_required_libraries(render_settings)
}
\arguments{
\item{render_settings}{the list containing available_libs (vector of files in the package)
and output_file (path of the rendered report, where directory "lib" is searched)}
}
\description{
Check if the files (e.g. javascript) required by the report are available in the main package
}
