% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{climate_narrative}
\alias{climate_narrative}
\title{Main funtion that runs the shiny app}
\usage{
climate_narrative(...)
}
\arguments{
\item{...}{Additional parameters to shiny server function}
}
\description{
Main funtion that runs the shiny app
}
