% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delete_empty_rows_and_columns}
\alias{delete_empty_rows_and_columns}
\title{Helper function - the name is self-explanatory}
\usage{
delete_empty_rows_and_columns(
  data,
  empty_strings = list("", "N/A"),
  ignore_cols = c()
)
}
\arguments{
\item{data}{matrix or data.frame, possibly containing missing value}

\item{empty_strings}{which strings should be considered as "empty"?}

\item{ignore_cols}{indices of columns which should be ignored for emptiness check}
}
\value{
input object without rows and columns where all entries are empty (i.e. NA or ""). NULL when the input is all empty
}
\description{
Helper function - the name is self-explanatory
}
