% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Called on package attach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Not used}

\item{pkgname}{Not used}
}
\description{
It turned out that it is necessary to define the resource path at this point, otherwise
it was not possible to deploy the app properly. John Adcock claims this is a well known issue
and a typical solution
}
