% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_cell}
\alias{exposure_grid_cell}
\title{Produce the layout of questionnaire tabs (cell, row, whole table)}
\usage{
exposure_grid_cell(
  exposure_item,
  prefix,
  tooltip_text = "",
  dev = FALSE,
  width = NULL
)
}
\arguments{
\item{exposure_item}{the name (appended to shiny input name)}

\item{prefix}{prefix that will be used for inputs id (to ensure they are unique)}

\item{tooltip_text}{text of the tooltip (blank by default)}

\item{dev}{development mode (default is FALSE and means "" option is initially selected
if set to TRUE then all options are selected as High to facilitate debugging}

\item{width}{passed to selectInput}
}
\description{
Produce the layout of questionnaire tabs (cell, row, whole table)
}
