% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_server}
\alias{exposure_grid_server}
\title{Produce a table of inputs with selectInput fields}
\usage{
exposure_grid_server(
  input,
  output,
  exposure_matrix,
  tooltip_matrix,
  label,
  dev = FALSE,
  width = NULL,
  transpose
)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{exposure_matrix}{Exposure matrix to show}

\item{tooltip_matrix}{Matrix of tooltips}

\item{label}{Label for grid}

\item{dev}{Are we in developement mode}

\item{width}{Width of each dropdown}

\item{transpose}{whether to transpose the grid (relative to exposure_matrix)}
}
\description{
Produce a table of inputs with selectInput fields
}
