% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_ui}
\alias{exposure_grid_ui}
\title{Grid table of inputs (everything happens in matching server function)}
\usage{
exposure_grid_ui(label)
}
\arguments{
\item{label}{the name of output table to look for}
}
\description{
Grid table of inputs (everything happens in matching server function)
}
