% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_yaml_section}
\alias{find_yaml_section}
\title{Parse the (subset of) YAML text and find the matching section}
\usage{
find_yaml_section(string, start, end, indentation, section_name)
}
\arguments{
\item{string}{YAML text to parse.}

\item{start}{First line of parsing (choose 1 to parse the text from the start)}

\item{end}{Last line of parsing (choose length(string) to parse the text to the end)}

\item{indentation}{The indentation of the whole block of text to strip first.}

\item{section_name}{The header to look for.}
}
\value{
List with the following named elements:
start (first line of section found, excluding the name),
end (last line of section found)
indentation (indentation applied within the section)
}
\description{
Parse the (subset of) YAML text and find the matching section
}
