% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_yaml_subsection}
\alias{find_yaml_subsection}
\title{Recursively applies find_yaml_section function to find the subsection in a nested structure}
\usage{
find_yaml_subsection(string, section_subsection)
}
\arguments{
\item{string}{String to parse.}

\item{section_subsection}{The vector of string. Its lenght represents level of nesting, each value is a name
of section at corresponding level (starting from the top level)}
}
\description{
Recursively applies find_yaml_section function to find the subsection in a nested structure
}
