% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_images}
\alias{format_images}
\title{Go through the markdown file, find all png images and scale down where relevant}
\usage{
format_images(filename, image_settings)
}
\arguments{
\item{filename}{markdown file to process}

\item{image_settings}{the list of necessary setting, comprising of:
\itemize{
\item target_width target width of images
default is 7 inches which roughly matches vertical A4 page with margins
\item target_width_units either "in" for inches of "\%"
\item fix_width if TRUE all images will be scaled exactly to target_width. otherwise,
only the larger images will be scaled down
\item min_pixels_to_rescale for fix_width=TRUE, pictures narrower than this number of pixels are not scaled up
this is to prevent ugly look of upscaled low resolution images
\item max_height maximum height of picture after rescaling. Another mechanism to prevent
too large upscaled pictures (in particular square or portrait layout)
}}
}
\value{
NULL, changes file specified as an argument in place
}
\description{
Go through the markdown file, find all png images and scale down where relevant
}
