% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generic_footer}
\alias{generic_footer}
\title{Produce a footer HTML text explaining materiality levels}
\usage{
generic_footer(
  asset_or_liability = c("asset", "liability"),
  is_asset_mananger = FALSE
)
}
\arguments{
\item{asset_or_liability}{see above}

\item{is_asset_mananger}{see above}
}
\description{
Depending on parameters the materiality is defined in relation to:
"premium income" (asset_or_liability="liability"),
"assets under management" (asset_or_liability="assets" and is_asset_manager=TRUE), or
"assets" (asset_or_liability="assets" and is_asset_manager=FALSE)
}
