% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generic_liability_footer}
\alias{generic_liability_footer}
\title{Produce a footer HTML text explaining materiality levels (liabilities)}
\usage{
generic_liability_footer()
}
\description{
Produce a footer HTML text explaining materiality levels (liabilities)
}
