% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_appendices}
\alias{get_appendices}
\title{Function to produce appendices (as references)}
\usage{
get_appendices(items, exposure_classes)
}
\arguments{
\item{items}{vector of items to get references for}

\item{exposure_classes}{List of exposure classes (sectors)}
}
\value{
markdown-formatted references section (h2)
}
\description{
Function to produce appendices (as references)
}
