% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_executive_summary}
\alias{get_executive_summary}
\title{Function to combine an executive summary from lower level functions}
\usage{
get_executive_summary(
  tabs,
  scenarios,
  exposure_classes,
  aggregated_inputs,
  inputs,
  scenario_no,
  layout = 1
)
}
\arguments{
\item{tabs}{List of tabs (required for summary of user inputs)}

\item{scenarios}{List of scenarios}

\item{exposure_classes}{List of sectors}

\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}

\item{scenario_no}{integer or vector of integers indicating which of the global$scenarios should be included}

\item{layout}{determines the structure of the output

1 for full three subsections (inputs, scenarios, exposures)
2 for one subsection (scenarios, including also exposures)}
}
\value{
string - executive summary text
}
\description{
Function to combine an executive summary from lower level functions
}
