% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_executive_summary_exposures}
\alias{get_executive_summary_exposures}
\title{One of the functions comprising the executive summary text}
\usage{
get_executive_summary_exposures(aggregated_inputs, inputs, scenario_no)
}
\arguments{
\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}

\item{scenario_no}{integer or vector of integers indicating which of the global$scenarios should be included}
}
\value{
string - executive summary text
}
\description{
One of the functions comprising the executive summary text
}
