% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_executive_summary_inputs}
\alias{get_executive_summary_inputs}
\title{One of the functions comprising the executive summary text}
\usage{
get_executive_summary_inputs(aggregated_inputs, inputs)
}
\arguments{
\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}
}
\value{
string - executive summary text
}
\description{
One of the functions comprising the executive summary text
}
