% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_exposure_appendix}
\alias{get_exposure_appendix}
\title{Produce appendix for a given item}
\usage{
get_exposure_appendix(item)
}
\arguments{
\item{item}{name of item for which appendix is to be produced}
}
\value{
markdown-formatted appendix section (h3)
}
\description{
Produce appendix for a given item
}
