% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_exposure_risk_description}
\alias{get_exposure_risk_description}
\title{Lower level report helper function responsible for single risk (transition/physical) description}
\usage{
get_exposure_risk_description(
  item,
  products,
  materiality,
  physical_or_transition,
  high_or_low
)
}
\arguments{
\item{item}{the exposure item (sector or equivalent abstraction class)}

\item{products}{a vector of products (within the item)}

\item{materiality}{if "high" more detailed description is given}

\item{physical_or_transition}{"physical" or "transition" (to select the risk description)}

\item{high_or_low}{may be "high", "low" or FALSE in which case empty string is returned}
}
\description{
Lower level report helper function responsible for single risk (transition/physical) description
}
