% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_exposure_test_description}
\alias{get_exposure_test_description}
\title{Karnan's request for easier change comparison - single sector}
\usage{
get_exposure_test_description(item)
}
\arguments{
\item{item}{sector name}
}
\description{
Karnan's request for easier change comparison - single sector
}
