% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_input_ids}
\alias{get_input_ids}
\title{Produce a table of input names for an input table}
\usage{
get_input_ids(exposure_matrix, label, tab = NULL)
}
\arguments{
\item{exposure_matrix}{Exposure matrix upon which the input table is based}

\item{label}{Label for grid}

\item{tab}{NULL by default. If given, the first two arguments are ignored and inferred from tab}
}
\description{
Produce a table of input names for an input table
}
