% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_input_values}
\alias{get_input_values}
\title{Produce a matrix of input values from a matrix of input names (simple double loop)}
\usage{
get_input_values(inputs, ids_matrix)
}
\arguments{
\item{inputs}{dataframe of (all) inputs to look for}

\item{ids_matrix}{matrix of ids}
}
\description{
Produce a matrix of input values from a matrix of input names (simple double loop)
}
