% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_or_null}
\alias{get_or_null}
\title{Get object by its name but return NULL (not error) if it does not exist}
\usage{
get_or_null(name)
}
\description{
Get object by its name but return NULL (not error) if it does not exist
}
