% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_references}
\alias{get_references}
\title{Function to produce references section (for all items)}
\usage{
get_references(items)
}
\arguments{
\item{items}{vector of items to get references for}
}
\value{
markdown-formatted references section (h2)
}
\description{
Function to produce references section (for all items)
}
