% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_report_contents}
\alias{get_report_contents}
\title{Function that generates a markdown content of the report}
\usage{
get_report_contents(aggregated_inputs, inputs, report_version)
}
\arguments{
\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}

\item{report_version}{enables different versions of the reports within a single code, see global file for possible choices and their meaning}
}
\value{
vector of string - executive summary text (3 items + 1 per scenario + 1 item at the end)
}
\description{
Function that generates a markdown content of the report
}
