% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_report_settings}
\alias{get_report_settings}
\title{Create a single list of settings from simple arguments}
\usage{
get_report_settings(
  content_files,
  output_file,
  md_file,
  file_format,
  report_version,
  dev,
  rep_type,
  inst_type,
  report_sector_selection,
  report_scenario_selection
)
}
\arguments{
\item{content_files}{List of necessary global lists with report contents}

\item{output_file}{Path and filenamename of report to write}

\item{md_file}{Path and filename of intermediate markdown file}

\item{file_format}{Currently either "html" or "rtf"}

\item{report_version}{Integer controlling the version of the code used in report generating functions}

\item{rep_type}{Either "inst" for institutional report or "sect" for sectoral report}

\item{inst_type}{Institution type (relevant for institutional report only)}

\item{report_sector_selection}{Input used to filter report contents}

\item{report_scenario_selection}{Input used to filter report contents}
}
\value{
list of lists
}
\description{
Create a single list of settings from simple arguments
}
