% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scenario_descriptions}
\alias{get_scenario_descriptions}
\title{Lower level report helper function responsible for single scenario description}
\usage{
get_scenario_descriptions(aggregated_table, type_inputs, scenario)
}
\arguments{
\item{aggregated_table}{table of aggregated inputs (per items)}

\item{type_inputs}{disaggregated inputs table (to list all the contributing rows)}

\item{scenario}{the list with the following fields:
name, description, is_scenario, transition, physical}
}
\description{
Lower level report helper function responsible for single scenario description
}
