% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scenario_no}
\alias{get_scenario_no}
\title{Helper function that translate the value of input field to the scenario number(s)}
\usage{
get_scenario_no(scenarios, report_scenario_selection, is_rtf)
}
\arguments{
\item{scenarios}{List of scenarios}

\item{report_scenario_selection}{(user-friendly) scenario name (or empty string)}

\item{is_rtf}{a flag that triggers several format specific settings:
\itemize{
\item TRUE: include non-scenario sections (e.g. intro)
\item FALSE: include the links to page top (note requires proper report_version as well)
}}
}
\value{
vector of integers
}
\description{
Helper function that translate the value of input field to the scenario number(s)
}
