% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_section_descriptions}
\alias{get_section_descriptions}
\title{Lower level report helper function responsible for single section description}
\usage{
get_section_descriptions(section, additional_pars = list())
}
\arguments{
\item{section}{Section name}

\item{additional_pars}{List of additional parameters that may be required by special content functions}
}
\description{
Lower level report helper function responsible for single section description
}
