% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_section_no}
\alias{get_section_no}
\title{Helper function that returns non-scenario section number(s)}
\usage{
get_section_no(sections, is_rtf, rep_type)
}
\arguments{
\item{sections}{List of non-scenario report sections}

\item{is_rtf}{a flag that triggers several format specific settings:
\itemize{
\item TRUE: include non-scenario sections (e.g. intro)
\item FALSE: include the links to page top
}}

\item{rep_type}{additional possibility to filter report sections, by default (NULL) no filtering}
}
\value{
vector of string - executive summary text (3 items + 1 per scenario + 1 item at the end)
}
\description{
Helper function that returns non-scenario section number(s)
}
