% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GreCAPTCHAv3Ui}
\alias{GreCAPTCHAv3Ui}
\title{Google captcha function (UI part)}
\usage{
GreCAPTCHAv3Ui(siteKey)
}
\arguments{
\item{siteKey}{the (non-secret) sitekey given by google}
}
\description{
based on
https://github.com/sarthi2395/shinygCAPTCHAv3/blob/master/R/shinygCAPTCHAv3.R
}
