% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heartbeat}
\alias{heartbeat}
\title{Heartbeat function (server part) to prevent app closing due to inactivity}
\usage{
heartbeat(input, output, session)
}
\arguments{
\item{input}{standard shiny server parameter}

\item{output}{standard shiny server parameter}

\item{session}{standard shiny server parameter}
}
\description{
Heartbeat function (server part) to prevent app closing due to inactivity
}
