% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heartbeat_footer}
\alias{heartbeat_footer}
\title{Heartbeat function (ui part) to prevent app closing due to inactivity}
\usage{
heartbeat_footer()
}
\description{
Heartbeat function (ui part) to prevent app closing due to inactivity
}
