% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{include_markdown_section}
\alias{include_markdown_section}
\title{Read the markdown section from relevant yaml file and save as output}
\usage{
include_markdown_section(output, output_name, section_name)
}
\arguments{
\item{output}{Shiny output}

\item{output_name}{Name of slot created in shiny output (to be used in UI part of shiny app)}

\item{section_name}{Name of yaml file within section directory read from
(description section only, other YAML fields are ignored)}
}
\description{
This function should be used on the server side.
Corresponding UI side call should be: uiOutput(output_name)
}
