% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{load_secrets}
\alias{load_secrets}
\title{Read the yaml file containing app settings}
\usage{
load_secrets(secrets_file = "secret.yml")
}
\arguments{
\item{secrets_file}{Path to the file (by default "secret.yml" in the current directory)}
}
\description{
Read the yaml file containing app settings
}
