% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paste_recursive}
\alias{paste_recursive}
\title{Helper function to merge lists of lists with the same (possibly nested) structure, ultimately containing strings}
\usage{
paste_recursive(list_1, list_2, name_1, name_2, sep = "\\n\\n")
}
\arguments{
\item{list_1}{list 1}

\item{list_2}{list 2}

\item{name_1}{first name (to be appended at the lowest level)}

\item{name_2}{second name (to be appended at the lowest level)}

\item{sep}{separator between names and list contents}
}
\description{
Helper function to merge lists of lists with the same (possibly nested) structure, ultimately containing strings
}
