% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{prepare_user_data}
\alias{prepare_user_data}
\title{Helper function creating necessary data in session$userData}
\usage{
prepare_user_data(session)
}
\arguments{
\item{session}{Shiny session.}
}
\description{
Helper function creating necessary data in session$userData
}
