% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_full_report}
\alias{produce_full_report}
\title{Function that writes a full report to (temporary) file}
\usage{
produce_full_report(report_contents, tempfile)
}
\arguments{
\item{report_contents}{the content to write}

\item{tempfile}{where to write the report}
}
\description{
this is necessary as markdown::render takes file as an argument
not used at the moment, but do not delete - will be sent by email probably
}
