% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_add.R
\name{produce_report}
\alias{produce_report}
\title{The highest level function for report production. Takes only two arguments}
\usage{
produce_report(all_inputs, settings)
}
\arguments{
\item{all_inputs}{Table of user inputs}

\item{settings}{list of lists containing all necessary settings}
}
\value{
NULL, report produced to file
}
\description{
The highest level function for report production. Takes only two arguments
}
