% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_selective_report}
\alias{produce_selective_report}
\title{Function that writes a full report to (temporary) file}
\usage{
produce_selective_report(
  report_contents,
  report_scenario_selection,
  include_common_sections,
  tempfile
)
}
\arguments{
\item{report_contents}{the content to write}

\item{report_scenario_selection}{(user-friendly) scenario name (or empty string)}

\item{include_common_sections}{whether to include non-scenario sections (e.g. intro)}

\item{tempfile}{where to write the report}
}
\description{
this is necessary as markdown::render takes file as an argument
}
