% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dir}
\alias{read_dir}
\title{Read all files from a directory as a named R list}
\usage{
read_dir(
  directory,
  file_format = "auto",
  in_package = TRUE,
  remove_special_characters_from_names = TRUE
)
}
\arguments{
\item{directory}{directory to parse}

\item{file_format}{format of the files to determine the correct parser.
If "auto" (the default) the format will be guessed by extension of the first file}

\item{in_package}{bool. If true (the default) the directory will be looked for in the package directory}

\item{remove_special_characters_from_names}{bool. If true (the default) the name of the resulting list will be stripped
of special characters (in particular underscores)}
}
\description{
Handles yaml/csv/R files.
}
