% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recaptcha_js}
\alias{recaptcha_js}
\title{Google captcha function, actual javascript call.}
\usage{
recaptcha_js(site_key, action, field_id)
}
\arguments{
\item{site_key, }{Site key from google}

\item{action}{Name of action to test}

\item{field_id}{Field to trigger}
}
\description{
Based on
https://github.com/sarthi2395/shinygCAPTCHAv3/blob/master/R/shinygCAPTCHAv3.R
}
