% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recaptcha_ui}
\alias{recaptcha_ui}
\title{Google captcha function, UI part.}
\usage{
recaptcha_ui(site_key)
}
\arguments{
\item{site_key}{Site key from google}
}
\description{
based on
https://github.com/sarthi2395/shinygCAPTCHAv3/blob/master/R/shinygCAPTCHAv3.R
}
