% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_special_characters}
\alias{remove_special_characters}
\title{Remove special characters (in particular spaces) from a string}
\usage{
remove_special_characters(text, make_camelcase = TRUE)
}
\arguments{
\item{text}{string to convert}

\item{make_camelcase}{bool. If true (the default), the output will be CamelCase}
}
\description{
Additionally (optionally) make it camelcase.
}
