% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_yaml_subsection}
\alias{replace_yaml_subsection}
\title{Parse the yaml file, replace the selected piece and save (overwrite)}
\usage{
replace_yaml_subsection(yaml_file_location, section_subsection, new_text)
}
\arguments{
\item{yaml_file_location}{location of YAML file to update.}

\item{section_subsection}{vector of strings - recursively go down the YAML structure to find the desired location.}

\item{new_text}{new content.}
}
\description{
Parse the yaml file, replace the selected piece and save (overwrite)
}
