% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{restore_spaces}
\alias{restore_spaces}
\title{Format (camelcase) string in order to look better in final output (spaces, capitalisation)}
\usage{
restore_spaces(camelcase)
}
\arguments{
\item{camelcase}{the string to convert}
}
\description{
Format (camelcase) string in order to look better in final output (spaces, capitalisation)
}
