% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rtf_center_images}
\alias{rtf_center_images}
\title{Center images in RTF}
\usage{
rtf_center_images(filename)
}
\arguments{
\item{filename}{name of file to convert}
}
\value{
NULL, changes file specified as an argument in place
}
\description{
By default pandoc rtf aligns pictures left, to center them (consistently with HTML)
manual intervention is required
}
