% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rtf_fix_table_of_contents}
\alias{rtf_fix_table_of_contents}
\title{Fix the table of contents in the RTF file}
\usage{
rtf_fix_table_of_contents(filename, dev)
}
\arguments{
\item{filename}{name of file to convert}

\item{dev}{developmen mode flag (used only to decide whether a warning is raised)}
}
\value{
NULL, changes file specified as an argument in place
}
\description{
ToC in pandoc output is not working out of the box. The ToC is a list of hyperlinks,
but there are no corresponding bookmarks in headers of respective sections
}
