% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rtf_postprocess}
\alias{rtf_postprocess}
\title{Fix the RTF file}
\usage{
rtf_postprocess(filename, dev)
}
\arguments{
\item{filename}{name of file to convert}

\item{dev}{developmen mode flag (used only to decide whether a warning is raised)}
}
\value{
NULL, changes file specified as an argument in place
}
\description{
It is not possible to set some options using markdown syntax. Also, there seems to be a bug
in pandoc which breaks the TOC (links are not working)
}
