% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Main function that runs the shiny app.}
\usage{
run_shiny_app(secrets_file = "secret.yml", ...)
}
\arguments{
\item{secrets_file}{Location of yaml file containing secrets}

\item{...}{Additional parameters to shiny server function}
}
\description{
Main function that runs the shiny app.
}
