% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{separate_toc}
\alias{separate_toc}
\title{Helper function to separate the table of contents from the main document}
\usage{
separate_toc(filename, file_contents = NULL)
}
\arguments{
\item{filename}{HTML file to parse}

\item{file_contents}{(optionally) conents of the HTML file (if already parsed, to avoid duplicating this work)}
}
\value{
NULL. The HTML file given as argument filename is updated as file without ToC,
which is saved as a separate HTML in the same directory as filename (with the name appended with "_toc")
}
\description{
Helper function to separate the table of contents from the main document
}
