% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{string_add_spaces_to_make_equal_lines}
\alias{string_add_spaces_to_make_equal_lines}
\title{Add spaces to a string so that it can be split into blocks of exactly the same length}
\usage{
string_add_spaces_to_make_equal_lines(string, line_width)
}
\arguments{
\item{string}{the string to convert}

\item{line_width}{the desired number of characters per line}
}
\description{
without breaking words
}
