% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{string_break_line_with_spaces}
\alias{string_break_line_with_spaces}
\title{Insert spaces to the string so that line has exactly given number of characters.}
\usage{
string_break_line_with_spaces(string, line_width, location, n_char = 1)
}
\arguments{
\item{string}{String to break}

\item{line_width}{Width of line}

\item{location}{Position of the (first character of) the break in the string}

\item{n_char}{Number of characters that are to be replaced with a (manual space-based) linebreak}
}
\description{
Insert spaces to the string so that line has exactly given number of characters.
}
