% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tab_name_to_number}
\alias{tab_name_to_number}
\title{Helper function to enable using names to refer tabs instead of numbers}
\usage{
tab_name_to_number(tab_name)
}
\arguments{
\item{tab_name}{name of the tab to convert}
}
\description{
Helper function to enable using names to refer tabs instead of numbers
}
