% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_to_markdown}
\alias{table_to_markdown}
\title{A simple version of function to produce markdown tables from R table}
\usage{
table_to_markdown(table, additional_spaces = 3, dot_to_space = TRUE)
}
\arguments{
\item{table}{R data frame}

\item{additional_spaces}{Additional space to add}

\item{dot_to_space}{Convert dots in titles to spaces}
}
\description{
It does not handle multiline cells.
}
