% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_delete.R
\name{write_report_to_file}
\alias{write_report_to_file}
\title{Function that writes a (full or selective) report to a (temporary) file}
\usage{
write_report_to_file(report_contents, tempfile, fix_image_width = FALSE)
}
\arguments{
\item{report_contents}{the content to write}

\item{tempfile}{where to write the report}

\item{fix_image_width}{whether to set all the images to fixed width}
}
\value{
NULL, output is a file as specified in the argument
}
\description{
this is necessary as markdown::render takes file as an argument
}
