% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate_quantitative_to_qualitative_materiality}
\alias{aggregate_quantitative_to_qualitative_materiality}
\title{Aggregate multiple numerical materiality exposures into a single, qualitative}
\usage{
aggregate_quantitative_to_qualitative_materiality(x)
}
\description{
Aggregate multiple numerical materiality exposures into a single, qualitative
}
