% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ensure_images_fit_page}
\alias{ensure_images_fit_page}
\title{Go through the markdown file, find all png images and scale down where relevant}
\usage{
ensure_images_fit_page(filename, max_width_inch = 7)
}
\arguments{
\item{filename}{the path to file to convert}

\item{max_width_inch}{maximum width of image in inches (wider will be scaled down to this value)}
}
\description{
Go through the markdown file, find all png images and scale down where relevant
}
