% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_cell}
\alias{exposure_grid_cell}
\title{Produce the layout of questionnaire tabs (cell, row, whole table)}
\usage{
exposure_grid_cell(
  exposure_item,
  prefix,
  tooltip_text = "",
  dev = FALSE,
  width = NULL
)
}
\description{
Produce the layout of questionnaire tabs (cell, row, whole table)
}
