% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_server}
\alias{exposure_grid_server}
\title{Produce a table of inputs with selectInput fields}
\usage{
exposure_grid_server(
  input,
  output,
  exposure_matrix,
  tooltip_matrix,
  label,
  dev = FALSE,
  width = NULL
)
}
\description{
Produce a table of inputs with selectInput fields
}
