% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exposure_grid_server}
\alias{exposure_grid_server}
\title{Produce a table of inputs with selectInput fields}
\usage{
exposure_grid_server(
  input,
  output,
  exposure_matrix,
  label,
  dev = FALSE,
  width = NULL
)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{exposure_matrix}{Exposure matrix to show}

\item{label}{Label for grid}

\item{dev}{Are we in developement mode}

\item{width}{Width of each dropdown}
}
\description{
Produce a table of inputs with selectInput fields
}
