% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generic_footer}
\alias{generic_footer}
\title{Produce a footer HTML text explaining materiality levels.}
\usage{
generic_footer(asset_or_liability, is_asset_mananger = FALSE)
}
\arguments{
\item{asset_or_liability}{A string with the type of exposure page}

\item{is_asset_mananger}{Is this display for an asset manager}
}
\description{
Produce a footer HTML text explaining materiality levels.
}
