% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generic_footer}
\alias{generic_footer}
\title{Produce a footer HTML text explaining materiality levels}
\usage{
generic_footer(
  asset_or_liability = c("asset", "liability"),
  is_asset_mananger = FALSE
)
}
\description{
Produce a footer HTML text explaining materiality levels
}
