% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_executive_summary}
\alias{get_executive_summary}
\title{Function to combine an executive summary from lower level functions}
\usage{
get_executive_summary(aggregated_inputs, inputs)
}
\arguments{
\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}
}
\value{
vector of string - executive summary text (3 items + 1 per scenario + 1 item at the end)
}
\description{
Function to combine an executive summary from lower level functions
}
