% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_executive_summary_scenarios}
\alias{get_executive_summary_scenarios}
\title{One of the functions comprising the executive summary text}
\usage{
get_executive_summary_scenarios(
  scenarios,
  exposure_classes,
  aggregated_inputs,
  inputs,
  scenario_no,
  exposure_exec = FALSE
)
}
\arguments{
\item{scenarios}{List of scenarios}

\item{exposure_classes}{List of sectors}

\item{aggregated_inputs}{data frame of aggregated inputs (implemented in the reactive expression)}

\item{inputs}{data frame of all inputs (implemented in the reactive expression)}

\item{scenario_no}{integer or vector of integers indicating which of the global$scenarios should be included}

\item{exposure_exec}{bool whether to include a short description for the sector
(applicable in single sector context only)}
}
\value{
string - executive summary text
}
\description{
One of the functions comprising the executive summary text
}
