% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_exposure_risk_description}
\alias{get_exposure_risk_description}
\title{Lower level report helper function responsible for single risk
(transition/physical) description.}
\usage{
get_exposure_risk_description(
  item,
  products,
  materiality,
  physical_or_transition,
  high_or_low,
  include_oneliner = FALSE
)
}
\arguments{
\item{item}{Exposure class name}

\item{products}{Product name}

\item{materiality}{Materiality of item}

\item{physical_or_transition}{Type of scenario}

\item{high_or_low}{Is scenario high or low}

\item{include_oneliner}{FALSE by default, normally oneliner is for executive summary}
}
\description{
Lower level report helper function responsible for single risk
(transition/physical) description.
}
