% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_exposure_risk_description}
\alias{get_exposure_risk_description}
\title{Lower level report helper function responsible for single risk (transition/physical) description}
\usage{
get_exposure_risk_description(
  item,
  products,
  materiality,
  physical_or_transition,
  high_or_low
)
}
\description{
Lower level report helper function responsible for single risk (transition/physical) description
}
