% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_references}
\alias{get_references}
\title{Function to produce references section (for all items)}
\usage{
get_references(aggregated_table, type_inputs)
}
\arguments{
\item{aggregated_table}{aggregated inputs}

\item{type_inputs}{disaggregated inputs}
}
\value{
markdown-formatted references section (h2)
}
\description{
Function to produce references section (for all items)
}
