% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_report_contents}
\alias{get_report_contents}
\title{Gets either a standard or test report contents}
\usage{
get_report_contents(content_files, inputs, content_settings)
}
\arguments{
\item{content_files}{list of all yaml content files}

\item{inputs}{user dropdown choices}

\item{content_settings}{all configuration options (see get_standard_report_content for details)}
}
\description{
Gets either a standard or test report contents
}
