% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scenario_descriptions}
\alias{get_scenario_descriptions}
\title{Lower level report helper function responsible for single scenario description}
\usage{
get_scenario_descriptions(
  aggregated_table_by_item,
  aggregated_table_by_group,
  type_inputs,
  scenario,
  exposure_classes,
  include_exposures
)
}
\arguments{
\item{aggregated_table_by_item}{Table of all possible items}

\item{aggregated_table_by_group}{As above but less granular}

\item{type_inputs}{Drop box items}

\item{scenario}{Scenario name}

\item{exposure_classes}{List of exposure classes (sectors) from which the texts are extracted}

\item{include_exposures}{Flag whether to include tables with contributing exposures}
}
\description{
Lower level report helper function responsible for single scenario description
}
