% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scenario_descriptions}
\alias{get_scenario_descriptions}
\title{Lower level report helper function responsible for single scenario description}
\usage{
get_scenario_descriptions(aggregated_table, type_inputs, scenario)
}
\arguments{
\item{aggregated_table}{Table of all possible items}

\item{type_inputs}{Drop box items}

\item{scenario}{Scenario name}
}
\description{
Lower level report helper function responsible for single scenario description
}
