% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_scenario_no}
\alias{get_scenario_no}
\title{Helper function that translate the value of input field to the scenario number(s)}
\usage{
get_scenario_no(report_scenario_selection, is_rtf)
}
\arguments{
\item{report_scenario_selection}{value of the input}

\item{is_rtf}{bool whether to include sections that are for RTF only}
}
\value{
vector of integers
}
\description{
Helper function that translate the value of input field to the scenario number(s)
}
