% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_section_no}
\alias{get_section_no}
\title{Helper function that returns non-scenario section number(s)}
\usage{
get_section_no(is_rtf)
}
\arguments{
\item{is_rtf}{bool whether to include sections that are for RTF only}
}
\value{
vector of integers
}
\description{
Helper function that returns non-scenario section number(s)
}
