% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_test_report}
\alias{get_test_report}
\title{Karnan's request for easier change comparison - loop over all sectors}
\usage{
get_test_report()
}
\description{
Karnan's request for easier change comparison - loop over all sectors
}
