% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GreCAPTCHAv3Server}
\alias{GreCAPTCHAv3Server}
\title{Google captcha function (server part)}
\usage{
GreCAPTCHAv3Server(secretKey, reCaptchaResponse)
}
\description{
based on
https://github.com/sarthi2395/shinygCAPTCHAv3/blob/master/R/shinygCAPTCHAv3.R
}
