% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{html_postprocess}
\alias{html_postprocess}
\title{Process the HTML file manually to correct some known issues}
\usage{
html_postprocess(file)
}
\arguments{
\item{file}{the location of HTML file}
}
\description{
Includes fixing image links (so they work in shiny)
and adding arrow to all headers (report version 5 only)
}
