% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_report_}
\alias{produce_report_}
\title{Function where actual report production takes place}
\usage{
produce_report_(all_inputs, settings)
}
\value{
NULL, report produced to file
}
\description{
Function where actual report production takes place
}
