% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_selective_report}
\alias{produce_selective_report}
\title{Function that writes a full report to (temporary) file}
\usage{
produce_selective_report(
  report_contents,
  report_version,
  report_scenario_selection,
  is_rtf,
  tempfile
)
}
\arguments{
\item{report_contents}{the content to write}

\item{report_version}{variable that can be used to control the multiple report versions in a single code}

\item{report_scenario_selection}{(user-friendly) scenario name (or empty string)}

\item{is_rtf}{a flag that triggers several format specific settings:
\itemize{
\item TRUE: include non-scenario sections (e.g. intro)
\item FALSE: include the links to page top (note requires proper report_version as well)
}}

\item{tempfile}{where to write the report}
}
\value{
NULL, output is a file as specified in the argument
}
\description{
this is necessary as markdown::render takes file as an argument
}
