% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{produce_tooltip_matrix}
\alias{produce_tooltip_matrix}
\title{Produce a matrix of tooltips (strings) by concatenating column-specific (if any)
and product-specific text (if any)}
\usage{
produce_tooltip_matrix(exposure_matrix)
}
\arguments{
\item{exposure_matrix}{the matrix of exposures in the specific format
Its first column is name, second is product, the others contain exposure names or blank cells}
}
\description{
Produce a matrix of tooltips (strings) by concatenating column-specific (if any)
and product-specific text (if any)
}
