% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dir}
\alias{read_dir}
\title{Read all files from a directory as a named R list.
Handles yaml/csv/R files.}
\usage{
read_dir(
  directory,
  file_format = "auto",
  in_package = system.file(directory, package = "climate.narrative") != "",
  remove_special_characters_from_names = TRUE
)
}
\arguments{
\item{directory}{Directory to process}

\item{file_format}{File format to interpret files as. One of yml, csv, r or auto}

\item{in_package}{Is the directory in a package}

\item{remove_special_characters_from_names}{Should we remove special characters}
}
\description{
Read all files from a directory as a named R list.
Handles yaml/csv/R files.
}
