% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dir}
\alias{read_dir}
\title{Read all files from a directory as a named R list}
\usage{
read_dir(
  directory,
  file_format = "auto",
  in_package = TRUE,
  remove_special_characters_from_names = TRUE
)
}
\description{
Handles yaml/csv/R files.
}
