% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{string_break_line_with_spaces}
\alias{string_break_line_with_spaces}
\title{Insert spaces to the string so that line has exactly given number of characters}
\usage{
string_break_line_with_spaces(string, line_width, location, n_char = 1)
}
\description{
Insert spaces to the string so that line has exactly given number of characters
}
