% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{string_format_lines}
\alias{string_format_lines}
\title{Format the string by appending spaces so it exactly fills the lines of given length}
\usage{
string_format_lines(string, col_width)
}
\description{
additionally, if the string contains at least one "br" tag
format output as a bulleted list
}
