% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_to_markdown}
\alias{table_to_markdown}
\title{A simple version of function to produce markdown tables from R table
(does not handle multiline cells)}
\usage{
table_to_markdown(table, additional_spaces = 3, dot_to_space = TRUE)
}
\description{
A simple version of function to produce markdown tables from R table
(does not handle multiline cells)
}
