% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_to_markdown_multiline}
\alias{table_to_markdown_multiline}
\title{Produce a markdown table out of R data frame}
\usage{
table_to_markdown_multiline(table, dot_to_space = TRUE, col_widths = NULL)
}
\arguments{
\item{table}{the data.frame to be converted}

\item{dot_to_space}{if TRUE (the default) the dots in table headers
will be replaced with space (useful as R table headers cannot contain spaces)}

\item{col_widths}{desired width of columns in the markdown file.
Note: it should translate roughly proportionately to column widths in the rendered doc}
}
\description{
Create a markdown table, allowing multiline cell entries (lines need to be separated by br tag)
The function splits the text automatically and adds spaces to match the desired column width
without breaking words
}
